Bug fix
Fixed database models

+++ api/v0.0.1/model/list.js
@@ -12,11 +12,13 @@ const Response = require("../../../core/response");
  */
 const checkRights = async (token, listId) => {
 	const account = await accountModel.auth(token);
-	const list = await models.List.findOne({where: {id: listId}, raw: true})
+	//const list = await models.List.findOne({where: {id: listId}, raw: true})
 
-	if(list && list.accoundId === account.id){
+	const ourMember = await models.List.findOne({where: {accountId: account.id, id: listId}, raw: true});
+	if(ourMember){
 		return true;
 	}
+
 	return Response.error(400, "Access is denied");
 }
 
@@ -28,12 +30,14 @@ const checkRights = async (token, listId) => {
  * @param description
  */
 const create = async ({token, deskId, name, description}) => {
-	await accountModel.auth(token);
+	const account = await accountModel.auth(token);
+	const accountId = account.id;
 	await deskModel.checkRights(token, deskId);
-	models.List.create({name, description, deskId}).then((res) => {
+	models.List.create({name, description, deskId, accountId}).then((res) => {
 		return Response.success({listId: res.id});
 	});
-	return Response.error(500, "Some wrong");
+	// Жопа, already sent exception
+	//return Response.error(500, "Some wrong");
 }
 
 /**
@@ -43,6 +47,8 @@ const create = async ({token, deskId, name, description}) => {
  * @returns {Promise<void>}
  */
 const get = async ({token, id}) => {
+	// Избегаем попытки повторной отправки ответа
+	let isSent = false;
 	// Getting list
 	const list = await models.List.findOne({where: {id: id}, raw: true});
 	// If he fined
@@ -51,7 +57,11 @@ const get = async ({token, id}) => {
 		if(await deskModel.getMember(token, list.deskId) !== true) return Response.error(500, "Access denied");
 		return Response.success({list: list});
 	}
-	return Response.error(500, "List not found");
+	else
+	{
+		return Response.error(500, "List not found");
+	}
+
 }
 
 /**
@@ -79,12 +89,25 @@ const edit = async ({token, id, name, description}) => {
  * @returns {Promise<void>}
  */
 const remove = async ({token, id}) => {
-	// Check our rights for it
-	await checkRights(token, id);
-	models.List.destroy({where: {id: id}}).then(() => {
-		return Response.success();
-	});
-	return Response.error(500, "Some wrong");
+	// Избегаем попытки повторной отправки ответа
+	let isSent = false;
+	let access = await checkRights(token, id);
+	if(access === true)
+	{
+
+		try {
+			models.List.destroy({where: {id: id}}).then(() => {
+				if (!isSent) {
+					return Response.success();
+				}
+
+			});
+		} catch (e) {
+			isSent = true;
+			return Response.error(500, "Some wrong");
+
+		}
+	}
 }
 
 module.exports = { checkRights, create, get, edit, remove }
